pyproject.toml
[project]
name = "scholub"
version = "0.1.0"
description = "Backend of scholub"
readme = "README.md"
requires-python = "==3.13.3"
dependencies = [
  "apscheduler==3.11.0",
  "argon2-cffi==23.1.0",
  "arxiv==2.2.0",
  "authlib==1.5.2",
  "fastapi[standard]==0.115.12",
  "gel==3.1.0",
  "itsdangerous==2.2.0",
  "pydantic==2.11.4",
  "pyjwt==2.10.1",
  "pytest-asyncio==0.26.0",
  "python-dotenv==1.1.0",
]

[dependency-groups]
dev = [
  "pytest==8.3.5",
]

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.lint.per-file-ignores]
"queries/*" = ["F403", "I"]

[tool.basedpyright]
exclude = ["queries/*"]

